requirements.txt
flask
tensorflow-cpu
gunicorn
pillow
numpy
mahotas